refactor: forbidden erro export
