+++ src/errors/index.js
@@ -1,4 +1,5 @@
 export * from './bad-request.error';
+export * from './forbidden.error'
 export * from './http-error';
 export * from './internal-server.error';
 export * from './not-found.error';
